token auth added

+++ src/components/WorkoutDetails.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import { ToastContainer, toast } from 'react-toastify';
 
 // date fns
@@ -7,11 +8,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
 
-  const {dispatch} = useWorkoutsContext()
+  const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async() => {
+    if(!user) {
+      return
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
     })
     const json = await response.json()
